+++ fugitive-darkness-provider/fugitive-darkness-provider-docx/src/main/java/com/zer0s2m/fugitivedarkness/provider/docx/ContainerInfoFileContent.java
@@ -1,0 +1,14 @@
+package com.zer0s2m.fugitivedarkness.provider.docx;
+
+import java.util.Collection;
+
+public record ContainerInfoFileContent(
+
+        int lineNumber,
+
+        String line,
+
+        Collection<ContainerInfoFileContentMatch> matchers
+
+) {
+}