+++ fugitive-darkness-provider/fugitive-darkness-provider-docx/src/main/java/com/zer0s2m/fugitivedarkness/provider/docx/ContainerInfoSearchDocxFile.java
@@ -1,0 +1,14 @@
+package com.zer0s2m.fugitivedarkness.provider.docx;
+
+import java.util.List;
+
+public record ContainerInfoSearchDocxFile(
+
+        String file,
+
+        String pattern,
+
+        List<ContainerInfoFileContent> found
+
+) {
+}