+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/GitRepoProviderInfo.java
@@ -1,0 +1,16 @@
+package com.zer0s2m.fugitivedarkness.provider.git;
+
+import java.util.Map;
+
+public record GitRepoProviderInfo(
+
+        String host,
+
+        String scheme,
+
+        String path,
+
+        Map<String, String> headers
+
+) {
+}