+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/SearchEngineGitException.java
@@ -1,0 +1,7 @@
+package com.zer0s2m.fugitivedarkness.provider.git;
+
+/**
+ * The main exception is for the search engine.
+ */
+public abstract class SearchEngineGitException extends Exception {
+}